# Conflicts:
#	TheColony/Main/GameObjectManager.cpp

+++ TheColony/Main/Camera.cpp
@@ -72,7 +72,7 @@ XMFLOAT3 Camera::ScreenToWorldSpace(XMFLOAT3 _screenPos) const
 	XMVECTOR screenVec = XMVectorSet(_screenPos.x, _screenPos.y, _screenPos.z, 1);
 	RECT viewPort = Application::GetInstance()->GetWindowRect();
 	XMMATRIX worldMat = gameObject->GetComponent<Transform>()->GetLocalMatrix();
-	XMVECTOR unProj = XMVector3Unproject(screenVec, viewPort.left, viewPort.top, viewPort.right - viewPort.left, viewPort.bottom - viewPort.top, nearPlane, farPlane, projectionMatrix, viewMatrix, worldMat);
+	XMVECTOR unProj = XMVector3Unproject(screenVec, (float)viewPort.left, (float)viewPort.top, (float)viewPort.right - (float)viewPort.left, (float)viewPort.bottom - (float)viewPort.top, nearPlane, farPlane, projectionMatrix, viewMatrix, worldMat);
 	XMMATRIX tra = XMMatrixTranslation(unProj.m128_f32[0], unProj.m128_f32[1], unProj.m128_f32[2]) * worldMat;
 	unProj = tra.r[3];
 	return XMFLOAT3(unProj.m128_f32[0], unProj.m128_f32[1], unProj.m128_f32[2]);
